+++ src/main/java/com/konkuk/Eodikase/domain/auth/service/AuthService.java
@@ -1,9 +1,9 @@
 package com.konkuk.Eodikase.domain.auth.service;
 
-import com.konkuk.Eodikase.dto.request.AuthLoginRequest;
-import com.konkuk.Eodikase.dto.request.KakaoLoginRequest;
-import com.konkuk.Eodikase.dto.response.OAuthTokenResponse;
-import com.konkuk.Eodikase.dto.response.TokenResponse;
+import com.konkuk.Eodikase.dto.request.auth.AuthLoginRequest;
+import com.konkuk.Eodikase.dto.request.auth.KakaoLoginRequest;
+import com.konkuk.Eodikase.dto.response.auth.OAuthTokenResponse;
+import com.konkuk.Eodikase.dto.response.auth.TokenResponse;
 import com.konkuk.Eodikase.domain.member.entity.Member;
 import com.konkuk.Eodikase.domain.member.entity.MemberPlatform;
 import com.konkuk.Eodikase.domain.member.entity.MemberStatus;
