+++ src/main/java/com/konkuk/Eodikase/dto/request/auth/AuthLoginRequest.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.request;
+package com.konkuk.Eodikase.dto.request.auth;
 
 import lombok.*;
