+++ src/main/java/com/konkuk/Eodikase/dto/request/data/FilteredCourseDataCountRequest.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.request;
+package com.konkuk.Eodikase.dto.request.data;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
