+++ src/main/java/com/konkuk/Eodikase/dto/request/member/MemberProfileUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.request;
+package com.konkuk.Eodikase.dto.request.member;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
