+++ src/main/java/com/konkuk/Eodikase/dto/request/member/PasswordVerifyRequest.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.request;
+package com.konkuk.Eodikase.dto.request.member;
 
 import lombok.*;
