+++ src/main/java/com/konkuk/Eodikase/dto/response/auth/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.response;
+package com.konkuk.Eodikase.dto.response.auth;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
