+++ src/main/java/com/konkuk/Eodikase/dto/response/data/FilteredCourseDataByRadiusResponse.java
@@ -1,28 +1,38 @@
-package com.konkuk.Eodikase.dto.response;
+package com.konkuk.Eodikase.dto.response.data;
 
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-public class CourseDataDetailInfoResponse {
+public class FilteredCourseDataByRadiusResponse {
+
     private Long id;
+
     private String name;
+
     private String category;
+
     private String location;
+
     private String phoneNumber;
+
     private Float scoreByNaver;
-    private String href;
+
     private String operatingTime;
+
     private int reviewCount;
+
     private String imageUrl;
-    private String img1;
-    private String img2;
-    private String img3;
+
     private double lat;
+
     private double lng;
 
-    public CourseDataDetailInfoResponse(Long id, String name, String category, String location, String phoneNumber,
-                                        Float scoreByNaver, String href, String operatingTime, int reviewCount,
-                                        String imageUrl, String img1, String img2, String img3, double lat, double lng
+    public FilteredCourseDataByRadiusResponse(
+            Long id, String name, String category, String location, String phoneNumber, Float scoreByNaver,
+            String operatingTime, int reviewCount, String imageUrl, double lat, double lng
     ) {
         this.id = id;
         this.name = name;
@@ -30,13 +40,9 @@ public class CourseDataDetailInfoResponse {
         this.location = location;
         this.phoneNumber = phoneNumber;
         this.scoreByNaver = scoreByNaver;
-        this.href = href;
         this.operatingTime = operatingTime;
         this.reviewCount = reviewCount;
         this.imageUrl = imageUrl;
-        this.img1 = img1;
-        this.img2 = img2;
-        this.img3 = img3;
         this.lat = lat;
         this.lng = lng;
     }
