+++ src/main/java/com/konkuk/Eodikase/dto/response/member/MyPageResponse.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.response;
+package com.konkuk.Eodikase.dto.response.member;
 
 import lombok.*;
