+++ src/main/java/com/konkuk/Eodikase/dto/response/member/OAuthMemberSignUpResponse.java
@@ -1,4 +1,4 @@
-package com.konkuk.Eodikase.dto.response;
+package com.konkuk.Eodikase.dto.response.member;
 
 import lombok.*;
